Updated PSS client to be consistent with other clients
